Extract Items Filter to own component
